+++ src/components/info-message/InfoMessage.tsx
